// frontend/src/app/core/api/model/time-slot-definition-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TimeSlotDefinitionRequest { 
    code: string;
    label: string;
    startTime: string;
    endTime: string;
    orderIndex: number;
    timeWindowValid?: boolean;
}